R7.3: split vector/module into build/outputs/lifecycle/check
Facade mod.rs keeps the crate-facing surface (pub(crate) use build::*);
no public path changes. Producer orchestration of the checked pipeline
(build_verified_vector_module{,_with_evidence}, fallback evidence, and
the producer-side admission guard
reject_cross_loop_delay_read_transports) lives in build.rs; output
materialization in outputs.rs; final assembly, lifecycle functions,
prototypes and compute drivers in lifecycle.rs; final-module
verification (sound-field immutability, mutable-table/UI-write
attribution, readonly-table stores, final shape) in check.rs.

Plan §4.8 note: module has no standalone checker entry point —
verify_final_module is called only from the producer's terminal step.
No admission guard changed paths in this split;
reject_cross_loop_delay_read_transports keeps its single call site in
build.rs and its existing rejection tests.

Validation: transform 387 lib tests, vector_mode oracle 35/35,
workspace tests green, workspace clippy (+1.97.0) clean, release
compiler build clean.
