+++ porting/journal/2026-07-20.md
@@ -152,3 +152,18 @@ modules on both sides of the plan artifact. Validation: transform 385,
 - Validation lib SANS cfg(test) incluse cette fois (build release
   compiler) après le leçon de R7.1.
 - Gates : transform 387, clippy 0, fmt vert, release compiler vert.
+
+## R7.3 — scission de `vector/module` — R7 complet
+
+- `module/mod.rs` (1 297 lignes) scindé en `build.rs` (355, orchestration
+  producteur + garde d'admission `reject_cross_loop_delay_read_transports`
+  + évidence de fallback), `outputs.rs` (138), `lifecycle.rs` (349,
+  assemblage final/prototypes/drivers), `check.rs` (455, vérifications du
+  module final), façade `mod.rs` (25, `pub(crate) use build::*`).
+- §4.8 (cas dégénéré comme lower) : pas d'entrée checker autonome —
+  `verify_final_module` n'est appelé que par l'étape terminale du
+  producteur. Aucune garde n'a changé de chemin ;
+  `reject_cross_loop_delay_read_transports` garde son site d'appel unique
+  dans build.rs et ses tests existants.
+- Gates : transform 387, vector_mode 35/35, workspace tests verts,
+  clippy workspace 0, release compiler vert.
